+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,3 +1,6 @@
+## v5.1.0-nightly-20260305080720 (Mar 05, 2026)
+## v5.1.0-dev (Mar 05, 2026)
+## v3.17.0-nightly-20260305080720 (Mar 05, 2026)
 ## v5.1.0-dev
 - Add Node 24 support. Drop Node 16 support. [#3652](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3652)
 - Add Shopper Consents API support [#3674](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3674)